chore: unify to provider error
