+++ hdp/src/provider/envelope/evm/provider.rs
@@ -1,8 +1,9 @@
 use crate::{
     primitives::processed_types::mmr::MMRMeta,
     provider::{
-        config::ProviderConfig, AccountProofsResult, AsyncResult, HeaderProofsResult,
-        ProofProvider, StorageProofsResult, TxProofsResult, TxReceiptProofsResult,
+        config::ProviderConfig, envelope::error::ProviderError, AccountProofsResult, AsyncResult,
+        HeaderProofsResult, ProofProvider, StorageProofsResult, TxProofsResult,
+        TxReceiptProofsResult,
     },
 };
 use alloy::{
@@ -18,46 +19,14 @@ use std::{
     collections::{HashMap, HashSet},
     time::Instant,
 };
-use thiserror::Error;
 use tracing::info;
 
 use crate::{
-    provider::indexer::{Indexer, IndexerError},
+    provider::indexer::Indexer,
     provider::types::{FetchedTransactionProof, FetchedTransactionReceiptProof},
 };
 
-use super::rpc::{RpcProvider, RpcProviderError};
-
-/// Error from [`EvmProvider`]
-#[derive(Error, Debug)]
-pub enum ProviderError {
-    /// Error when the query is invalid
-    #[error("Transaction index out of bound: requested index: {0}, length: {1}")]
-    OutOfBoundRequestError(u64, u64),
-
-    /// Error when the MMR meta is mismatched among range of requested blocks
-    #[error("MMR meta mismatch among range of requested blocks")]
-    MismatchedMMRMeta,
-
-    /// Error when the MMR is not found
-    #[error("MMR not found")]
-    MmrNotFound,
-
-    /// Error from the [`Indexer`]
-    #[error("Failed from indexer")]
-    IndexerError(#[from] IndexerError),
-
-    /// Error from [`RpcProvider`]
-    #[error("Failed to get proofs: {0}")]
-    RpcProviderError(#[from] RpcProviderError),
-
-    /// Error from [`eth_trie_proofs`]
-    #[error("EthTrieError: {0}")]
-    EthTrieError(#[from] eth_trie_proofs::EthTrieError),
-
-    #[error("Fetch key error: {0}")]
-    FetchKeyError(String),
-}
+use super::rpc::RpcProvider;
 
 /// EVM provider
 ///
